謎解き部屋を改良
